requirements.txt
streamlit
anthropic
PyPDF2
python-docx
pyperclip